requirements.txt
aiogram==2.25.1
openai==0.27.8
python-dotenv==1.0.0
APScheduler==3.10.1
pytz
python-dateutil
